refactor: [BE] redis구독 폴더 위치 game내로 변경
현재 game내에서만 사용중이므로
