+++ BE/src/game/game.module.ts
@@ -8,11 +8,11 @@ import { GameRoomService } from './service/game.room.service';
 import { QuizCacheService } from './service/quiz.cache.service';
 import { QuizSetModule } from '../quiz-set/quiz-set.module';
 import { QuizSetService } from '../quiz-set/service/quiz-set.service';
-import { ScoringSubscriber } from '../common/redis/subscribers/scoring.subscriber';
-import { TimerSubscriber } from '../common/redis/subscribers/timer.subscriber';
-import { RoomSubscriber } from '../common/redis/subscribers/room.subscriber';
-import { PlayerSubscriber } from '../common/redis/subscribers/player.subscriber';
-import { RedisSubscriberService } from '../common/redis/redis-subscriber.service';
+import { RedisSubscriberService } from './redis/redis-subscriber.service';
+import { ScoringSubscriber } from './redis/subscribers/scoring.subscriber';
+import { TimerSubscriber } from './redis/subscribers/timer.subscriber';
+import { RoomSubscriber } from './redis/subscribers/room.subscriber';
+import { PlayerSubscriber } from './redis/subscribers/player.subscriber';
 
 @Module({
   imports: [RedisModule, QuizSetModule],
